refactor: migrations
