+++ imports/ts-node.js
@@ -1,0 +1,3 @@
+import { register } from 'node:module'
+import { pathToFileURL } from 'node:url'
+register('ts-node/esm', pathToFileURL('./'))
