+++ scripts/drop-database.ts
