+++ tsconfig.json
@@ -4,5 +4,5 @@
     "outDir": "dist",
     "rootDir": "src"
   },
-  "include": ["@types", "src/**/*.ts"]
+  "include": ["@types", "src/**/*.ts", "scripts/**/*.ts"]
 }
